backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
google-genai>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
faster-whisper>=1.1.0
Pillow>=10.0.0
playwright>=1.49.0
